+++ src/Migration/Versioner/CannotParseVersionException.php
@@ -1,0 +1,25 @@
+<?php
+
+/*
+ * This file is part of the webmozart/json package.
+ *
+ * (c) Bernhard Schussek <bschussek@gmail.com>
+ *
+ * For the full copyright and license information, please view the LICENSE
+ * file that was distributed with this source code.
+ */
+
+namespace Webmozart\Json\Migration\Versioner;
+
+use RuntimeException;
+
+/**
+ * Thrown when a version cannot be parsed.
+ *
+ * @since  1.3
+ *
+ * @author Bernhard Schussek <bschussek@gmail.com>
+ */
+class CannotParseVersionException extends RuntimeException
+{
+}