+++ CHANGELOG.md
@@ -6,6 +6,8 @@ Changelog
  * added `JsonConverter` and `ConversionException`
  * added `MigratingConverter` to migrate JSON objects between different versions
  * added `ValidatingConverter` to validate converted JSON against schemas
+ * added `JsonVersioner` and implementations `VersionFieldVersioner` and
+   `SchemaFieldPatternVersioner`
  * added support for `$ref` to external schema
 
 * 1.2.2 (2016-01-14)